Always trim newlines from commands, even continuations
Previously, Net::IMAP::Server::Connection->parse_command trimmed
trailing newlines, meaning that parsed options would not have trailing
newlines.  However, continuation commands, such as for AUTHENTICATE,
were passed the unmodified input, including trailing \r\n.

As the \r\n is a property of the communication channel, and should not
be used for information transfer, remove them for consistency.  This
does not impact SASL implementations, however, as they are passed the
output of decode_base64, which is agnostic to the presence of trailing
whitespace.  Only third-party commands which use ->pending are affected
by this change.