Use recv() with MSG_PEEK to check for half-closed TCP sockets in test
After the client sends QUIT, the server closes its half of the
connection.  This leaves the connection in TCP status CLOSE_WAIT from
the point of view of the client.  Testing that the connection is
half-closed is slightly complicated; previously, we faked it by forcing
a write failure.

Use the low-level recv() function to examine if there is any data
waiting in the TCP buffer, of if there can ever be.  If the connection
is half-closed, it will return 0; otherwise, the server -> client
connection is understood to be still open.