Verify the SSL certificate used for testing
IO::Socket:SSL version 1.950 and above enforce SSL_VERIFY_PEER by
default, which ensures that the server has a signed certificate.  As
such, tests failed, since the certificate used in tests is self-signed.

Explicitly pass the signing CA (itself) when connecting or starting TLS;
this causes verification to succeed, which is essential to pass tests
with IO::Socket::SSL 1.950 or above.